+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add drag/drop and copy/paste of filenames and textures
 - add "reload"
+- better event handling in imagewindow
 
 # 2.6.3a, 7/1/24